AdminPanel/components cleaned up: put each component into their own folders.
