+++ app/adminpanel/components/ReviewTable/ReviewTable.tsx
@@ -19,7 +19,7 @@ export default function ReviewTable({
     return (
         <div className="bg-white p-6 rounded-lg shadow-md mb-6">
         <div className="flex justify-between items-center mb-4">
-            <h2 className="text-xl font-semibold">Reviews</h2>
+            <h2 className="text-xl font-semibold">Customer Reviews</h2>
             <button
             onClick={() => {
                 onAddClick("");
